Reduce default images-per-page setting of recount/rebuild fivefold
Originally, we were processing 200 images per recount/rebuild page, and
although the production forum I help administer doesn't even have that
many images, it took a little over 16 seconds for the recount/rebuild to
run on that forum, which seems excessive for a single page. I've reduced
it fivefold, to 40.