Fix file size bug; improve auto-orientation support; bump to v1.0.4
The `autorsz_fix_image_orientation()` function introduced in v1.0.3 did
not provide for:

1. The updating of the `$attachment['filesize']` entry, leading to an
incorrect value being stored in the database, and issues with the
display of images via `attachment.php` (partial rendering; failed
downloads), since the size sent to the browser by that script is pulled
out of the database rather than from the file in the filesystem.

2. Auto-orientation of the full set of eight orientation types.

This commit resolves both issues, and introduces code to correct for the
filesize bug of v1.0.3.